\name{cprint}
\alias{cprint}

\title{dump assignment}
\description{
  dump out an R assignemnt statement to the screen
}
\usage{
cprint(a)
}

\arguments{
  \item{a}{R object}
}

\value{
 
side effects
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}


\examples{

x = 10

cprint(x)

}


\keyword{misc}

