\name{helpcolors}
\alias{helpcolors}
\title{Help on Personal Color Palettes}
\description{
Give information on how to set up Personal Color Palettes
  
}
\usage{
helpcolors()
}

\value{
 Side effects
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{palette}
\examples{

helpcolors()




}
\keyword{misc}
