\name{pickcolors}
\alias{pickcolors}
\title{Pick a SYSTEM color}
\description{
  Pick a SYSTEM color
}
\usage{
pickcolors(COLLIST = colors(), BACK = "white")
}
\arguments{
  \item{COLLIST}{system colors}
  \item{BACK}{background for colors}
}

\value{
 List of colors
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{syscolors}

\keyword{misc}
