\name{rainbow.colors}
\alias{rainbow.colors}
\title{rainbow.colors}
\description{
  Color palette of n rainbow colors
}
\usage{
rainbow.colors(n)
}
\arguments{
  \item{n}{Nmber of colors desired}
}
\details{
 rainbow.colors is set to match other color palette selections like
 topo.colors, terrain.colors
}
\value{
  Character vector of n colors from the default rainbow palette.
}
\author{Jonathan M. Lees <jonathan.lees@unc.edu>}
\seealso{topo.colors, terrain.colors, palette}
\examples{
rainbow.colors(100)
}
\keyword{aplot}
