\name{wheelrgb}
\alias{wheelrgb}
\title{Plot a large color rectangle for color selection}
\description{
  Plot a large color rectangle for color selection
}
\usage{
wheelrgb(wloc, v, RY)
}
\arguments{
  \item{wloc}{output of locator}
  \item{v}{v, from hsv color scheme}
  \item{RY}{coordinates of meshgrid, output of  VVwheel}
}

\value{
  vector of colors
}

\author{Jonathan M. Lees<jonathan.lees@unc.edu>}

\seealso{colwheel,  VVwheel}

\keyword{misc}
